+++ executor/pipeline/accumulator.rs
